Add files via upload
This folder contains the initial dataset as produced by the python code, and the dataset in long format, which is the format used for Plot 1. These were uploaded separately to the R Markdown document to save time converting the data from the initial format to the long format. After some tests this seemed to save 30 seconds or so in loading the document.